+++ Sources/Main/Util/Private/URLRequest+RAT.swift
@@ -3,17 +3,13 @@ import Foundation
 extension URLRequest {
 
     /// Build a RAT http request
+    ///
     /// - Parameters:
     ///   - url: RAT url endpoint
     ///   - body: body data
     ///   - environmentBundle: type to get bundle data from
+    ///
     /// - Returns: url request
-    static func ratRequest(url: URL, body: Data, environmentBundle: EnvironmentBundle = Bundle.main) -> URLRequest {
-        return URLRequest(url: url, body: body, environmentBundle: environmentBundle)
-    }
-}
-
-fileprivate extension URLRequest {
     init(url: URL, body: Data, environmentBundle: EnvironmentBundle = Bundle.main) {
         self.init(url: url, cachePolicy: .reloadIgnoringCacheData, timeoutInterval: 30)
         httpShouldHandleCookies = environmentBundle.useDefaultSharedCookieStorage